+++ playbook/app/pb_kits/playbook/pb_background/docs/_background_image.md
@@ -1,0 +1,1 @@
+To add a lazyload on the background image simply use the `transition` prop and one of the three string options `"fade"`, `"blur"`, or `"scale"`.
