--- playbook/app/pb_kits/playbook/pb_background/docs/_background_transition.html.erb
@@ -1,60 +1,0 @@
-<%= pb_rails("flex") do %>
-  <%= pb_rails("flex/flex_item", props: {fixed_size: "250px"}) do %>
-    <%= pb_rails("select", props: {
-    id: "transition-dropdown",
-    label: "",
-    blank_selection: "Select a Transition...",
-    options: [
-        {
-          value: 'fade',
-        },
-        {
-          value: 'blur',
-        },
-        {
-          value: 'scale',
-        },
-    ]
-    }) %>
-  <% end %>
-    <%= pb_rails("flex/flex_item") do %>
-      <%= pb_rails("button", props: {classname: "button", text: "Load Image", margin_left: "sm", margin_top: "xs", disabled: true }) %>
-    <% end %>
-<% end %>
-<%= pb_rails("background",  props: { classname: "transition-image", alt: "colorful background", image_url: "https://images.unsplash.com/photo-1528459801416-a9e53bbf4e17?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1000&q=80", on_error: "this.style.display = 'none'" }) do %>
-  <%= pb_rails("flex", props: { horizontal: "center", orientation: "column" }) do %>
-      <%= pb_rails("flex/flex_item") do %>
-        <%= pb_rails("title", props: { dark: true, size: 1, text: "Background Kit Transition" })%>
-      <% end %>
-    <% end %>
-<% end%>    
-
-<script>
-    window.addEventListener('DOMContentLoaded', () => {
-        const button = document.querySelector(".button")
-        const transitionDropdown = document.querySelector("#transition-dropdown")
-        var option = ''
-
-        transitionDropdown.addEventListener('change', function() {
-          if (this.value === '') {
-            button.className = button.className + ' _disabled'
-          } else {
-            button.removeAttribute("disabled");
-            button.className = button.className.replace(/\_disabled/gi, '')
-            option = this.value
-          }
-        });
-
-        const updateTransition = (transition) => {
-          const imageTransition = document.querySelector('.transition-image')
-            imageTransition.removeAttribute("style")
-            imageTransition.classList.remove("fade", "blur", "scale", "lazyloaded")
-            imageTransition.classList.add(transition, "lazyload")
-            
-        }
-      
-        button.addEventListener('click', function() {
-          updateTransition(option)
-        })
-    });
-</script>
