--- playbook/app/pb_kits/playbook/pb_background/docs/_background_transition.md
@@ -1,1 +1,0 @@
-To add a transition, simply use the `transition` prop and one of the three string options `"fade"`, `"blur"`, or `"scale"`.
