+++ playbook/app/pb_kits/playbook/pb_background/docs/index.js
@@ -4,4 +4,3 @@ export { default as BackgroundGradient } from './_background_gradient.jsx'
 export { default as BackgroundImage } from './_background_image.jsx'
 export { default as BackgroundStatus } from './_background_status.jsx'
 export { default as BackgroundCategory } from './_background_category.jsx'
-export { default as BackgroundTransition } from './_background_transition.jsx'
